The link flags for python was incomplete
It did not query the list of libraries, so was not automatically linking against all of the system libraries that python might require.  This could create a race condition where, for example, libutil was not loaded (using Python 2.7 on Centos 6.7) because it was not declared as a dependency on the shared library.  

This change asks python for the list of libraries AND the list of linker flags and then correctly includes those when creating the shared object for the python plugin.